backend/requirements.txt
Django==4.2.7
PyMySQL==1.1.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
python-decouple==3.8